bedrock update 041105

+++ tools/cli/src/steps.rs
@@ -79,7 +79,7 @@ pub fn gen_userlib_import_code(
 pub use {}::{entry_ty_name} as Game;
 pub const APP_IDENTIFIER: &'static str = {userlib_name:?};
 pub const APP_TITLE: &'static str = {userlib_title:?};
-pub const APP_VERSION: (u32, u32, u32) = ({}, {}, {});",
+pub const APP_VERSION: (u16, u16, u16) = ({}, {}, {});",
         userlib_name.replace('-', "_"),
         userlib_version.major,
         userlib_version.minor,
